% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abbrev}
\alias{abbrev}
\title{abbrev}
\format{An object of class \code{"data"}.}
\usage{
abbrev
}
\description{
a data.table of postal abbreviations
}
\examples{
data(abbrev)
cat(paste(abbrev[1:10], sep=','))
}
\keyword{datasets}
