% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{address}
\alias{address}
\title{address}
\usage{
address(DT)
}
\arguments{
\item{DT}{Datatable containing street.num, street.direction.prefix, street.body, street.type, street.direction.suffix, street.unit, street.unit.type, and city state zip}
}
\description{
Conatenates exploded street information back into normal address.
}
\keyword{address}
\keyword{concatenate}
