% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{clean.city}
\alias{clean.city}
\title{clean.city}
\usage{
clean.city(cityStateZip, study.cities)
}
\arguments{
\item{cityStateZip}{character vector}

\item{study.cities}{character vector}
}
\description{
Extracts and cleans city from cityStateZip based on table lookup and matching.
}
\keyword{city,}
\keyword{clean}
