% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{clean.lic.num}
\alias{clean.lic.num}
\title{clean.lic.num}
\usage{
clean.lic.num(license.num)
}
\arguments{
\item{license.num}{character vector}
}
\description{
cleans marijuana license number
}
\keyword{city,}
\keyword{lic,}
\keyword{license,}
\keyword{num,}
\keyword{number}
