% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{clean.lic.type}
\alias{clean.lic.type}
\title{clean.lic.type}
\usage{
clean.lic.type(license.type)
}
\arguments{
\item{license.type}{character vector}
}
\description{
cleans marijuana license type for mmed locations
}
\keyword{city,}
\keyword{lic,}
\keyword{license,}
\keyword{type}
