% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{clean.name.key}
\alias{clean.name.key}
\title{clean.name.key}
\usage{
clean.name.key(name)
}
\arguments{
\item{name}{character vector with names}
}
\description{
create a unique name key
}
\keyword{anems}
\keyword{clean}
