% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{clean.name}
\alias{clean.name}
\title{clean.name}
\usage{
clean.name(name)
}
\arguments{
\item{name}{character vector with names}
}
\description{
Clean an d sanitize business (location) names
}
\keyword{anems}
\keyword{clean}
