% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{clean.special}
\alias{clean.special}
\title{clean.special}
\usage{
clean.special(DT, DT.name = "")
}
\arguments{
\item{DT}{datatable with address info, city, street.type, street.direction.prefix, street.body}

\item{DT.name}{string flag for c*rent}
}
\description{
Cleans up some custom strings in the c*star rent data.
}
\keyword{(Denver}
\keyword{any}
\keyword{clean}
\keyword{does}
\keyword{exceptions}
\keyword{for}
\keyword{have}
\keyword{north)}
\keyword{not}
\keyword{some}
\keyword{special}
\keyword{weird}
