% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{clean.state}
\alias{clean.state}
\title{clean.state}
\usage{
clean.state(state)
}
\arguments{
\item{state}{state input}
}
\description{
Just returns CO if state is missing.
}
\keyword{colorodo}
\keyword{silly,}
