% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{clean.street.num}
\alias{clean.street.num}
\title{clean.street.num}
\usage{
clean.street.num(street.num)
}
\arguments{
\item{street.num}{character vector of street.nums}
}
\description{
Cleans up street.num, makes sure that hypens are set correctly.
}
\keyword{street.num}
