% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{clean.street.regex}
\alias{clean.street.regex}
\title{clean.street.regex}
\usage{
clean.street.regex()
}
\description{
Builds regex used in clean.street (depends on data.table which had mclapply bug in R 3.4.0)
}
\keyword{clean}
\keyword{regex}
\keyword{street}
