% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{clean.street.unit}
\alias{clean.street.unit}
\title{clean.street.unit}
\usage{
clean.street.unit(vec.chars)
}
\arguments{
\item{vec.chars}{character vector of street.units. This function is very confusing}
}
\description{
Cleans up street.unit number.
}
\keyword{street.num}
