% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{explode.address}
\alias{explode.address}
\title{explode.address}
\usage{
explode.address(DT, study.cities)
}
\arguments{
\item{DT}{data.table needs to have fields street, cityStateZip}

\item{study.cities}{character vector}
}
\description{
Explodes a single string addresss using explode.street and explode.cityStateZip.
}
\keyword{street.num}
