% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{explode.cityStateZip}
\alias{explode.cityStateZip}
\title{explode.cityStateZip}
\usage{
explode.cityStateZip(address, study.cities)
}
\arguments{
\item{address}{charcater vector with address}

\item{study.cities}{character vector}
}
\description{
Explodes cityStateZip into city state and zip.
}
\keyword{street.num}
