% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{explode.street}
\alias{explode.street}
\title{explode.street}
\usage{
explode.street(street, abbrev)
}
\arguments{
\item{street}{charcater vector with address}

\item{abbrev}{data.table data shipped with package}
}
\description{
Explodes street, into component pieces for matching, return a data.table.
}
\keyword{explode,}
\keyword{street}
