% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{extract.state}
\alias{extract.state}
\title{extract.state}
\usage{
extract.state(cityStateZip)
}
\arguments{
\item{cityStateZip}{character vector with cityStateZip}
}
\description{
Extracts state from cityStateZip.
}
\keyword{extract,}
\keyword{state}
