% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{extract.zip}
\alias{extract.zip}
\title{extract.zip}
\usage{
extract.zip(cityStateZip)
}
\arguments{
\item{cityStateZip}{charcater vector with cityStateZip}
}
\description{
Extracts zipcode from cityStateZip.
}
\keyword{extract,}
\keyword{zip}
\keyword{zipcode,}
