% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{fill.missing.state}
\alias{fill.missing.state}
\title{fill.missing.state}
\usage{
fill.missing.state(x)
}
\arguments{
\item{x}{data.table consisting of address.id and zip}
}
\description{
Fills missing state by querying zips
}
\keyword{city,}
\keyword{fill}
\keyword{missing,}
\keyword{state}
\keyword{zip,}
