% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup.state}
\alias{lookup.state}
\title{lookup.state}
\format{An object of class \code{"data"}.}
\usage{
lookup.state
}
\description{
a data.table of state names and abbreviations
,
}
\examples{
data(lookup.state)
cat(paste(lookup.state[1:10], sep=','))
}
\keyword{datasets}
