% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup.zips}
\alias{lookup.zips}
\title{lookup.zips}
\format{An object of class \code{"data"}.}
\usage{
lookup.zips
}
\description{
a data.table of zips, citis and states
,
}
\examples{
data(lookup.zips)
cat(paste(lookup.zips[1:10], sep=','))
}
\keyword{datasets}
