% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{parcels.street.direction.suffix}
\alias{parcels.street.direction.suffix}
\title{parcels.street.direction.suffix}
\usage{
parcels.street.direction.suffix(street.explode, return.direction,
  abbrev.direction)
}
\arguments{
\item{street.explode}{data.table}

\item{return.direction}{no idea}

\item{abbrev.direction}{list of abbreviations to search for}
}
\description{
Helper function for explode.street.
}
\keyword{explode,}
\keyword{helper}
\keyword{street.explode,}
