% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{parcels.street.last.promote}
\alias{parcels.street.last.promote}
\title{parcels.street.last.promote}
\usage{
parcels.street.last.promote(update.log, street.explode)
}
\arguments{
\item{update.log}{logical vector}

\item{street.explode}{data.table}
}
\description{
Helper function for explode.street.
}
\keyword{explode,}
\keyword{helper}
\keyword{street.explode,}
