% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regex.colorado.cities}
\alias{regex.colorado.cities}
\title{regex.colorado.cities}
\format{An object of class \code{"character"}.}
\source{
\href{https://www.dropbox.com/s/ix2fdubcqulb0rj/hash.tables.R?dl=0}{hashtables.R}
}
\usage{
regex.colorado.cities
}
\description{
A regex string containing the same data in study.cities
}
\examples{
regex.colorado.cities
cat(regex.colorado.cities)
}
\references{
regex.colorado.cities
(\href{https://www.dropbox.com/s/ix2fdubcqulb0rj/hash.tables.R?dl=0}{hashtables.R})
}
\keyword{datasets}
