% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study.cities}
\alias{study.cities}
\title{study.cities}
\format{An object of class \code{"character"}.}
\source{
\href{https://www.dropbox.com/s/ix2fdubcqulb0rj/hash.tables.R?dl=0}{hashtables.R}
}
\usage{
study.cities
}
\description{
A list of front range cities in the study area
}
\examples{
study.cities
cat(paste(study.cities, sep=','))
}
\references{
study.cities
(\href{https://www.dropbox.com/s/ix2fdubcqulb0rj/hash.tables.R?dl=0}{hashtables.R})
}
\keyword{datasets}
