% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study.zips}
\alias{study.zips}
\title{study.zips}
\format{An object of class \code{"character"} with 205 observations.}
\source{
\href{https://www.dropbox.com/s/ix2fdubcqulb0rj/hash.tables.R?dl=0}{hashtables.R}
}
\usage{
study.zips
}
\description{
List of all zip codes in the study area
}
\examples{
study.zips
cat(min(study.zips), max(study.zips))
}
\references{
study.zips
(\href{https://www.dropbox.com/s/ix2fdubcqulb0rj/hash.tables.R?dl=0}{hashtables.R})
}
\keyword{datasets}
