% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{underscore}
\alias{underscore}
\title{underscore}
\usage{
underscore(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Convert spaces, dashes, and antying else to underscore.
}
\keyword{cound}
\keyword{explode,}
\keyword{helper,}
\keyword{street.explode,}
\keyword{words}
