% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{word.count}
\alias{word.count}
\title{word.count}
\usage{
word.count(street.cum)
}
\arguments{
\item{street.cum}{character vector}
}
\description{
Helper function for explode.street, sees how may words are in the holder street.cum, also binds some words together.
}
\keyword{cound}
\keyword{explode,}
\keyword{helper,}
\keyword{street.explode,}
\keyword{words}
