% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abbrev}
\alias{abbrev}
\title{abbrev}
\format{An object of class \code{"data.table"}.}
\source{
\href{https://www.dropbox.com/s/ix2fdubcqulb0rj/hash.tables.R?dl=0}{hashtables.R}
}
\usage{
abbrev
}
\description{
A dataset containing address abbreviations
}
\examples{
data(abbrev)
table(abbrev$class)
}
\references{
Postal service abbreviations created in hashtables.R
(\href{https://www.dropbox.com/s/ix2fdubcqulb0rj/hash.tables.R?dl=0}{hashtables.R})
}
\keyword{datasets}
