% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{clean.street}
\alias{clean.street}
\title{clean.street}
\usage{
clean.street(street, abbrev)
}
\arguments{
\item{street}{character vector of streets}

\item{abbrev}{data.table used for lookup}
}
\description{
Cleans up and standardizes street string. Needed for the street.explode function.
}
\keyword{street.explode}
