% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{explode.address}
\alias{explode.address}
\title{explode.address}
\usage{
explode.address(DT, lookup.address = NULL)
}
\arguments{
\item{DT}{data.table needs to have fields street, cityStateZip}

\item{lookup.address}{optional table from main data usually parcels.address}
}
\description{
Explodes a single string addresss using explode.street and explode.cityStateZip.
}
\keyword{street.num}
