% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{explode.cityStateZip}
\alias{explode.cityStateZip}
\title{explode.cityStateZip}
\usage{
explode.cityStateZip(DT)
}
\arguments{
\item{DT}{data.table with column name containg address or cityStateZip}
}
\description{
Explodes cityStateZip into city state and zip
}
\keyword{street.num}
