% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.string.R
\name{fill.missing.zip.city}
\alias{fill.missing.zip.city}
\title{fill.missing.zip.city}
\usage{
fill.missing.zip.city(DT)
}
\arguments{
\item{DT}{data.table with exploded zip city state}
}
\description{
Fills missing zip and city by looking up in parcels.address if it exists.
}
\keyword{city,}
\keyword{fill}
\keyword{missing,}
\keyword{state}
\keyword{zip,}
