% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study.cities}
\alias{study.cities}
\title{study.cities}
\format{An object of class \code{"character"}.}
\usage{
study.cities
}
\description{
A list of Colorado front range cities
}
\examples{
data(study.cities)
cat(paste(study.cities, sep=','))
}
\keyword{datasets}
