% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study.zips}
\alias{study.zips}
\title{study.zips}
\format{An object of class \code{"character"} with 205 observations.}
\usage{
study.zips
}
\description{
List of Colorado zip codes in the study area
}
\examples{
data(study.zips)
cat(min(study.zips), max(study.zips))
}
\references{
study.zips
}
\keyword{datasets}
